Tableau Workbook analysis and management.
A collection of tools that analyze, document, and support management of
Tableau Workbooks.

By design the tools are run from the command line. It's possible to
integrate the core TWB gem classes into automated systems.

By default the tools will access all of the Tableau Workbooks in the
current directory/folder. Normal and packaged Workbooks are supported.

Documentation is created in a variety of media including PDF, PNG, and
Markdown files.
In general, documentation is created on a per-Workbook basis.

Data is created as CSV files, each of which identifies the specific
entities subject to the analysis.
Multiple CSV files are required to relate different entities to one
another due to the nature of the relationships between entities and the
limitations of tabular data sets in capturing these relationships.
Individual CSV files can be linked together to provide the analytical
flexibility required in different scenarios.